pyproject.toml
[tool.poetry]
name = "mixer"
version = "0.1.0"
description = ""
authors = ["Moshe Sherman <mtshrmn@gmail.com>"]
license = "GPL"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
hidapi = "^0.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
